Updates references per serving v0.9.0 vendoring
 - Updates code and tests to refer the required data structure and method/function calls
 - duckv1beta1 -> duckv1 conditions
 - traffic percent value is now *int64
 - containerConcurrency validation method relocation

+++ pkg/kn/commands/service/configuration_edit_flags.go
@@ -238,7 +238,7 @@ func (p *ConfigurationEditFlags) Apply(
 	}
 
 	if cmd.Flags().Changed("concurrency-limit") {
-		err = servinglib.UpdateConcurrencyLimit(template, p.ConcurrencyLimit)
+		err = servinglib.UpdateConcurrencyLimit(template, int64(p.ConcurrencyLimit))
 		if err != nil {
 			return err
 		}
